Moved the decorator decorators.py, a dedicated file for different decorators.
